backend_django/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
itsdangerous>=2.2.0
cryptography>=44.0.0